Refactor component imports and remove console.log statement
