+++ src/app/classes/layout.tsx
@@ -1,4 +1,4 @@
-import Header from "@/components/common/Header";
+import Header from "@/components/layout/Header";
 import Sidebar from "@/components/layout/Sidebar";
 import { Metadata } from "next";
